Fix added arrow when export multicalibration in case of no axis text showed

+++ src/ui/panel_model/MultiCalibrationView.h
@@ -67,7 +67,7 @@ private:
 
     MultiCalibrationDrawing* mDrawing;
     QTextEdit* mTextArea;
-    qreal mButtonWidth;
+    int mButtonWidth;
 
     Button* mImageSaveBut;
     Button* mImageClipBut;
@@ -101,7 +101,7 @@ private:
     double mTminDisplay;
     double mTmaxDisplay;
     double mThreshold;
-    double mGraphHeight;
+    int mGraphHeight;
     QColor mCurveColor;
 
     QString mResultText;
